requirements.txt
langchain==0.1.14
langchain_community==0.0.31
langchain_core==0.1.40
langchain_groq==0.1.0
langchain_openai==0.1.1
llama_index==0.10.27
python-dotenv==1.0.1
streamlit==1.32.2
